Refactor UI and extend the functionality in LabGUIT2
Adjusted the layout of LabGUIT2 form, switching from a GridLayoutManager to a BorderLayout, and replaced the JSlider with a JProgressBar. Furthermore, added a JComboBox for color selection to the form.
In the LabGUIT2 class, introduced an event listener to the JSlider that sets the value to JProgressBar. Also, created a HashMap for color options and used it in filling the JComboBox with color options. Added an event listener to JComboBox to display the selected color's name in the lblInfo and changing lblInfo's text color. Changes aimed at enhancing the user experience and providing additional functionalities. Not Working at this stage.